unión microservicios con autenticación

+++ src/resolvers/alquiler_resolver.js
@@ -1,39 +1,64 @@
 const alquilerResolver = {
     Query: {
-        alquilerByUsername: (_, {username}, {dataSources}) => {
-            return dataSources.vehiculosAPI.getAlquilerByUsername(username);
+        alquilerByUsername: async(_, {username}, {dataSources, userIdToken}) => {
+            usernameToken = (await dataSources.authAPI.getUser(userIdToken)).username
+            if (username == usernameToken)
+                return await dataSources.vehiculosAPI.getAlquilerByUsername(username);
+            else
+                return null
         },
 
-        alquilerById: (_, {alquilerId}, {dataSources}) => {
-            return dataSources.vehiculosAPI.getAlquilerById(alquilerId);
+        alquilerById: async(_, {alquilerId}, {dataSources, userIdToken}) => {
+            userIsStaff = (await dataSources.authAPI.getUser(userIdToken)).is_staff
+            if (userIsStaff == true)
+                return dataSources.vehiculosAPI.getAlquilerById(alquilerId);  
+            else 
+                return null
         },
 
-        alquilerList: (_, {alquilerList}, {dataSources}) => {
-            return dataSources.vehiculosAPI.getAlquilerList(alquilerList);
+        alquilerList: async(_, {alquilerList}, {dataSources, userIdToken}) => {
+            userIsStaff = (await dataSources.authAPI.getUser(userIdToken)).is_staff
+            if (userIsStaff == true)
+                return dataSources.vehiculosAPI.getAlquilerList(alquilerList)
+            else 
+                return null
         }
     },
 
     Mutation: {
-        createAlquiler: (_, {alquiler}, {dataSources}) => {
-            return dataSources.vehiculosAPI.createAlquiler(alquiler);
+        createAlquiler: async(_, {alquiler}, {dataSources, userIdToken}) => {
+            usernameToken = (await dataSources.authAPI.getUser(userIdToken)).username
+                if (alquiler.username == usernameToken)
+                    return dataSources.vehiculosAPI.createAlquiler(alquiler)
+                else 
+                    return null
         },
 
-        actualizarAlquiler: (_, {alquilerId, alquilerInput}, {dataSources}) => {
-            const alquilerActInput = {
-                id: alquilerInput.id,
-                username: alquilerInput.username,
-                vehiculoId: alquilerInput.vehiculoId,
-                fechaInicio: alquilerInput.fechaInicio,
-                fechaFinal: alquilerInput.fechaFinal,
-                lugarEntrega: alquilerInput.lugarEntrega,
-                lugarRegreso: alquilerInput.lugarRegreso,
-                entregado: alquilerInput.entregado
+        actualizarAlquiler: async(_, {alquilerId, alquilerInput}, {dataSources, userIdToken}) => {
+            usernameToken = (await dataSources.authAPI.getUser(userIdToken)).username
+            if (alquilerInput.username == usernameToken){
+                const alquilerActInput = {
+                    id: alquilerInput.id,
+                    username: alquilerInput.username,
+                    vehiculoId: alquilerInput.vehiculoId,
+                    fechaInicio: alquilerInput.fechaInicio,
+                    fechaFinal: alquilerInput.fechaFinal,
+                    lugarEntrega: alquilerInput.lugarEntrega,
+                    lugarRegreso: alquilerInput.lugarRegreso,
+                    entregado: alquilerInput.entregado
+                }
+                return dataSources.vehiculosAPI.updateAlquiler(alquilerId, alquilerActInput);
             }
-            return dataSources.vehiculosAPI.updateAlquiler(alquilerId, alquilerActInput);
+            else 
+                return null
         },
 
-        eliminarAlquiler: (_, {alquilerId}, {dataSources}) => {
-            return dataSources.vehiculosAPI.deleteAlquiler(alquilerId);
+        eliminarAlquiler: async(_, {alquilerId}, {dataSources, userIdToken}) => {
+            userIsStaff = (await dataSources.authAPI.getUser(userIdToken)).is_staff
+            if (userIsStaff == true)
+                return dataSources.vehiculosAPI.deleteAlquiler(alquilerId);
+            else
+                return null;
         }
     }
 }
